feat(irq): 将 IRQ 号类型统一由 u8 改为 u16，支持更多中断线
- 修改 riscv 架构中 IRQ 号相关类型为 u16，扩展中断号范围
- 调整 IrqLine 结构体及相关函数，支持 u16 类型的 IRQ 号
- 修正 mmio 模块中分配 IRQ 号时的类型匹配
- 优化 irq.rs 中 IrqLine 的注册和释放逻辑，注释掉部分未使用代码
- 修正 plic 相关调用，避免类型转换错误
- 移除部分不必要的导入和宏派生，简化代码结构

+++ ostd/src/arch/riscv/irq.rs
@@ -3,7 +3,6 @@
 //! Interrupts.
 use alloc::{boxed::Box, fmt::Debug, sync::Arc, vec::Vec};
 use core::sync::atomic::{AtomicPtr, Ordering};
-use crate::trap::irq::IrqNum;
 
 use crossbeam_queue::ArrayQueue;
 use id_alloc::IdAlloc;
@@ -57,7 +56,7 @@ pub(crate) fn alloc_soft_irq() -> Option<IrqNum> {
         .unwrap()
         .lock()
         .alloc()
-        .map(|i| SOFTWARE_IRQ_BASE as u8 + i as u8)
+        .map(|i| SOFTWARE_IRQ_BASE as u16 + i as u16)
 }
 
 pub(crate) static IRQ_LIST: Once<Vec<IrqLine>> = Once::new();
@@ -70,7 +69,7 @@ pub(crate) fn init() {
     let mut list: Vec<IrqLine> = Vec::new();
     for i in 0..256 {
         list.push(IrqLine {
-            irq_num: i as u8,
+            irq_num: i as u16,
             callback_list: GuardSpinLock::new(Vec::new()),
         });
     }
@@ -183,7 +182,7 @@ impl IrqLine {
     }
 
     /// Get the IRQ number.
-    pub fn num(&self) -> u8 {
+    pub fn num(&self) -> u16 {
         self.irq_num
     }
